Move Module and SkelModule to different locations
- `base.module` is now located in `module`
- `base.skelmodule` is now located in `prototypes.skelmodule`
