+++ core/render/vi/__init__.py
@@ -1,10 +1,9 @@
 # noinspection PyUnresolvedReferences
 from viur.core.render.vi.user import UserRender as user  # this import must exist!
 from viur.core.render.json.default import DefaultRender, CustomJsonEncoder
-from viur.core import conf, exposed, securitykey, utils, errors
+from viur.core import Module, conf, exposed, securitykey, utils, errors
 from viur.core.utils import currentRequest, currentLanguage
 from viur.core.skeleton import SkeletonInstance
-from viur.core.base.module import Module
 import datetime
 import json
